Firebase android setup Done
