+++ lib/mobiliy_app.dart
@@ -1,0 +1,18 @@
+import 'package:flutter/material.dart';
+
+class MobiliaApp extends StatelessWidget {
+  const MobiliaApp({super.key});
+
+  // This widget is the root of your application.
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'MOBILIA',
+      theme: ThemeData(
+        colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
+        useMaterial3: true,
+      ),
+      home: const Scaffold(),
+    );
+  }
+}
